Revert "Feat/org lookup (#162)"
This reverts commit c7f48f4b7e72afd9bd53ae259b4dd43830e439f8.
